algorithm updates - working towards being able to setup strikecraft to a schedule
