+++ code/cmd/xmltree/algorithm/ionshields.go
@@ -44,7 +44,7 @@ func (j *Job) applyIonShields() (err error) {
 	}
 
 	// apply stats for each component
-	err = j.ApplyAll(components)
+	err = j.ApplyComponentAll(components)
 
 	return
 }
